# Struct alternative to LogEventProperty; optimizations