--- a/Setup.ps1
+++ b/Setup.ps1
@@ -2,10 +2,8 @@ $ErrorActionPreference = "Stop"
 
 $RequiredDotnetVersion =  $(cat ./global.json | convertfrom-json).sdk.version
 
-if ($(dotnet --version) -ne $RequiredDotnetVersion) {
-	Invoke-WebRequest "https://dot.net/v1/dotnet-install.ps1" -OutFile "./build/installcli.ps1"
-	& ./build/installcli.ps1 -InstallDir "$pwd/.dotnetcli" -NoPath -Version $RequiredDotnetVersion
-	if ($LASTEXITCODE) { exit 1 }
-}
+Invoke-WebRequest "https://dot.net/v1/dotnet-install.ps1" -OutFile "./build/installcli.ps1"
+& ./build/installcli.ps1 -InstallDir "$pwd/.dotnetcli" -NoPath -Version $RequiredDotnetVersion
+if ($LASTEXITCODE) { exit 1 }
 
 $env:Path = "$pwd/.dotnetcli;$env:Path"