--- a/build.sh
+++ b/build.sh
@@ -12,13 +12,14 @@ echo "🤖 Attempting to build..."
 for path in src/**/*.csproj; do
     dotnet build -f netstandard1.0 -c Release ${path}
     dotnet build -f netstandard1.3 -c Release ${path}
+    dotnet build -f netstandard2.0 -c Release ${path}
 done
 
 echo "🤖 Running tests..."
 for path in test/*.Tests/*.csproj; do
-    dotnet test -f netcoreapp2.0  -c Release ${path}
+    dotnet test -f netcoreapp2.2  -c Release ${path}
 done
 
 for path in test/*.PerformanceTests/*.PerformanceTests.csproj; do
-    dotnet build -f netcoreapp2.0  -c Release ${path}
+    dotnet build -f netcoreapp2.2  -c Release ${path}
 done