--- a/src/Serilog/Core/Sinks/SecondaryLoggerSink.cs
+++ b/src/Serilog/Core/Sinks/SecondaryLoggerSink.cs
@@ -40,13 +40,7 @@ namespace Serilog.Core.Sinks
         {
             if (logEvent == null) throw new ArgumentNullException(nameof(logEvent));
 
-            var copy = new LogEvent(
-                logEvent.Timestamp,
-                logEvent.Level,
-                logEvent.Exception,
-                logEvent.MessageTemplate,
-                logEvent.Properties.Select(p => new LogEventProperty(p.Key, p.Value)));
-
+            var copy = logEvent.Copy();
             _logger.Write(copy);
         }