… appveyor.yml
… global.json
… src/Serilog/Events/LogEventProperty.cs
… test/Serilog.PerformanceTests/AllocationsBenchmark.cs
… test/Serilog.PerformanceTests/BindingBenchmark.cs
… test/Serilog.PerformanceTests/Harness.cs
… test/Serilog.PerformanceTests/LevelControlBenchmark.cs
… test/Serilog.PerformanceTests/LogContextEnrichmentBenchmark.cs
… test/Serilog.PerformanceTests/MessageTemplateCacheBenchmark/MessageTemplateCacheBenchmark_Cached.cs
… test/Serilog.PerformanceTests/MessageTemplateCacheBenchmark/MessageTemplateCacheBenchmark_Leaking.cs
… test/Serilog.PerformanceTests/MessageTemplateParsingBenchmark.cs
… test/Serilog.PerformanceTests/NestedLoggerCreationBenchmark.cs
… test/Serilog.PerformanceTests/NestedLoggerLatencyBenchmark.cs
… test/Serilog.PerformanceTests/PipelineBenchmark.cs
… test/Serilog.PerformanceTests/Serilog.PerformanceTests.csproj
… test/Serilog.Tests/Context/LogContextTests.cs
… test/Serilog.Tests/Core/LogEventPropertyCapturingTests.cs
… test/Serilog.Tests/Serilog.Tests.csproj
… test/Serilog.UwpTests/Serilog.UwpTests.csproj