This replaces some internal usage of the class `LogEventProperty` with a new, very similar, struct `EventProperty`.

The changes reduce allocations in message template binding/event construction, and in `ForContext(name, value)`, by a very modest amount.

Along the way, a couple of small opportunistic changes - some `IEnumerable<T>` over arrays was replaced with direct array passing, and copying of events (i.e. for `WriteTo.Logger()`) was refactored to allow the new event's properties dictionary to be constructed with the correct capacity.

``` ini
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17763.503 (1809/October2018Update/Redstone5)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=2.2.204
  [Host] : .NET Core 2.2.5 (CoreCLR 4.6.27617.05, CoreFX 4.6.27618.01), 64bit RyuJIT
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3416.0
  Core   : .NET Core 2.2.5 (CoreCLR 4.6.27617.05, CoreFX 4.6.27618.01), 64bit RyuJIT
```

### Pipeline benchmark (.NET Core 2.2)

This shows 56 fewer bytes allocated (12%) when binding/constructing a log event with one property. The gain here will be in proportion to the number of properties bound from the message template.

The difference is consistent across .NET Core and Framework.

Before:

|       Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| EmitLogEvent | 431.6 ns | 6.900 ns | 6.454 ns | 0.1369 |     - |     - |     432 B |

After:

|       Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| EmitLogEvent | 416.2 ns | 4.975 ns | 4.653 ns | 0.1192 |     - |     - |     376 B |

### Remaining work

Some more benchmarking is needed. I ran the allocations benchmark suite but apart from in the simple logging case that matches the pipeline benchmark above, differences were mostly noise, buried  by the other work being done.

There also remains a regression in `Logger.BindMessageTemplate()` due to the need to present `LogEventProperty`s on the API surface.