@nblumhardt 2019-05-27
## Nested logger creation (.NET Core 2.2)

Removed an allocation (15% fewer bytes on the most common two cases); added a small `string` hot path optimization in `PropertyValueConverter`, reorder to avoid TLS and a `GetType()` call during property value conversion. Nice little speed-up.

Before:

|           Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| ForContextString | 109.5 ns | 1.3622 ns | 1.2742 ns | 0.0483 |     - |     - |     152 B |
|   ForContextType | 144.7 ns | 1.1916 ns | 1.1146 ns | 0.0482 |     - |     - |     152 B |
|    ForContextInt | 100.0 ns | 0.7645 ns | 0.7151 ns | 0.0559 |     - |     - |     176 B |

After:

|           Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |----------:|----------:|----------:|-------:|------:|------:|----------:|
| ForContextString |  64.56 ns | 0.9045 ns | 0.8461 ns | 0.0407 |     - |     - |     128 B |
|   ForContextType | 111.04 ns | 1.2752 ns | 1.0648 ns | 0.0407 |     - |     - |     128 B |
|    ForContextInt | 103.95 ns | 0.9308 ns | 0.8251 ns | 0.0482 |     - |     - |     152 B |