@nblumhardt 2019-05-27
## Binding benchmark (.NET Core 2.2)

The PR introduces a regression in `Logger.BindMessageTemplate()`; the method is not used internally, and doesn't appear to be used widely in the ecosystem (I checked [_Serilog.Extensions.Logging_](https://github.com/serilog/serilog-extensions-logging/blob/dev/src/Serilog.Extensions.Logging/Extensions/Logging/SerilogLogger.cs#L50) - one of the possible consumers mentioned in #746, but it's using `BindProperty()` instead).

To quantify:

Before:

|   Method |      Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |----------:|----------:|----------:|------:|--------:|-------:|------:|------:|----------:|
| BindZero |  59.31 ns | 0.6348 ns | 0.5627 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|  BindOne | 149.77 ns | 2.4367 ns | 2.0347 ns |  2.53 |    0.05 | 0.0279 |     - |     - |      88 B |
| BindFive | 461.22 ns | 4.4476 ns | 4.1603 ns |  7.78 |    0.09 | 0.1092 |     - |     - |     344 B |

After:

|   Method |      Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |----------:|----------:|----------:|------:|--------:|-------:|------:|------:|----------:|
| BindZero |  65.32 ns | 0.7988 ns | 0.7081 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|  BindOne | 223.89 ns | 0.9839 ns | 0.9203 ns |  3.43 |    0.04 | 0.0455 |     - |     - |     144 B |
| BindFive | 594.10 ns | 5.8526 ns | 5.4746 ns |  9.10 |    0.13 | 0.1364 |     - |     - |     432 B |