@nblumhardt 2019-05-27
I think this is ready to go, but I can't figure out why the UWP tests are failing on the newer SDK (not running a version of VS that suppports UWP projects :-))

Anyone have any ideas? 😊