@nblumhardt 2019-05-28
😓 ... going .. to ... get .. this ... building ... on new SDK!