@nblumhardt 2019-05-28
👏 building! Anyone keen to sanity check this? :-)