@nblumhardt 2019-05-28
(Will need a squash/merge.)

The churn here is due to wanting to use newer C# features - `readonly` and `in` structs in this case. Good opportunity to bring the toolchain forward, anyway. No breaking changes 🤞