@nblumhardt 2019-05-29
If all's quiet I'll shoot this through to `dev`, folks (I think it's a safe win...) 😁